# api/openapi.yaml
openapi: 3.1.0
info:
  title: StackPilot API
  version: 0.0.0
  description: StackPilot local control-plane API contract.
servers:
  - url: /
security:
  - bearerAuth: []
  - sessionAuth: []
paths:
...
  /api/v1/systems/{systemId}/start:
    parameters:
      - $ref: "#/components/parameters/SystemId"
    post:
      operationId: startSystem
      summary: Queue the Phase 1B root-service start operation
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartSystemRequest"
      responses:
        "202":
          description: The operation was queued or replayed idempotently.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationRef"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableManifest"
  /api/v1/systems/{systemId}/stop:
    parameters:
      - $ref: "#/components/parameters/SystemId"
    post:
      operationId: stopSystem
      summary: Queue a non-cancellable Phase 1B root-service stop operation
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StopSystemRequest"
      responses:
        "202":
          description: The stop operation was queued or replayed idempotently.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationRef"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/systems/{systemId}/restart:
    parameters:
      - $ref: "#/components/parameters/SystemId"
    post:
      operationId: restartSystem
      summary: Stop the active runtime and fresh-start the latest valid manifest
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartSystemRequest"
      responses:
        "202":
          description: The restart operation was queued or replayed idempotently.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationRef"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableManifest"
  /api/v1/services/{systemId}/{serviceId}/restart:
    parameters:
      - $ref: "#/components/parameters/SystemId"
      - $ref: "#/components/parameters/ServiceId"
    post:
      operationId: restartService
      summary: Restart one service and its transitive downstream closure
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StopSystemRequest"
      responses:
        "202":
          description: The service restart operation was queued or replayed idempotently.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationRef"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/systems/{systemId}/status:
    parameters:
      - $ref: "#/components/parameters/SystemId"
      - $ref: "#/components/parameters/WorkspaceIdQuery"
    get:
      operationId: getSystemRuntimeStatus
      summary: Read the server-computed active runtime state
      responses:
        "200":
          description: Current runtime status, or stopped when no active instance exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemRuntimeStatus"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/operations/{operationId}:
    parameters:
      - name: operationId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/OperationId"
    get:
      operationId: getOperation
      summary: Read an operation and its ordered structured steps
      responses:
        "200":
          description: Persisted operation snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/operations:
    get:
      operationId: listOperations
      summary: List the newest persisted operations and structured steps
      parameters:
        - $ref: "#/components/parameters/WorkspaceIdQuery"
      responses:
        "200":
          description: At most 50 operations in newest-first order.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    maxItems: 50
                    items:
                      $ref: "#/components/schemas/Operation"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/operations/{operationId}/cancel:
    parameters:
      - name: operationId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/OperationId"
    post:
      operationId: cancelOperation
      summary: Record cooperative cancellation for a cancellable operation
      responses:
        "202":
          description: Cancellation was recorded; terminal compensation may still be running.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationRef"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Long-lived local token read only from OS-protected storage by CLI clients.
    sessionAuth:
      type: apiKey
      in: cookie
      name: stackpilot_session
  parameters:
...
    WorkspaceIdQuery:
      name: workspaceId
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/WorkspaceId"
    SystemId:
      name: systemId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ManifestId"
    ServiceId:
      name: serviceId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ManifestId"
...
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - details
        - traceId
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        operationId:
          type:
            - string
            - "null"
          pattern: "^op_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
        traceId:
          type: string
          pattern: "^tr_[0-9a-f]{32}$"
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/Error"
    OperationId:
      type: string
      pattern: "^op_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
    StartSystemRequest:
      type: object
      additionalProperties: false
      properties:
        workspaceId:
          $ref: "#/components/schemas/WorkspaceId"
        portOverrides:
          type: object
          additionalProperties:
            type: integer
            minimum: 1024
            maximum: 65535
        failurePolicy:
          type: object
          additionalProperties: false
          properties:
            failFast:
              type: boolean
            cleanupOnFailure:
              type: boolean
            keepReadyServices:
              type: boolean
    StopSystemRequest:
      type: object
      additionalProperties: false
      properties:
        workspaceId:
          $ref: "#/components/schemas/WorkspaceId"
    OperationRef:
      type: object
      additionalProperties: false
      required: [operationId, state]
      properties:
        operationId:
          $ref: "#/components/schemas/OperationId"
        state:
          type: string
          enum: [queued, running, cancelling, succeeded, failed, cancelled]
    OperationStep:
      type: object
      additionalProperties: false
      required: [number, key, state, attempt]
      properties:
        number:
          type: integer
          minimum: 1
        key:
          type: string
        state:
          type: string
          enum: [pending, running, succeeded, failed, skipped, cancelled]
        attempt:
          type: integer
          minimum: 0
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        durationMs:
          type: integer
          minimum: 0
        errorCode:
          type: string
        detailRef:
          type: string
    Operation:
      type: object
      additionalProperties: false
      required: [id, workspaceId, systemId, type, state, cancellable, createdAt, steps]
      properties:
        id:
          $ref: "#/components/schemas/OperationId"
        workspaceId:
          $ref: "#/components/schemas/WorkspaceId"
        systemId:
          $ref: "#/components/schemas/ManifestId"
        type:
          type: string
          enum: [start, stop, restart, service-restart, port-plan, refresh, analyze]
        state:
          type: string
          enum: [queued, running, cancelling, succeeded, failed, cancelled]
        cancellable:
          type: boolean
        cancelRequestedAt:
          type: string
          format: date-time
        errorCode:
          type: string
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        durationMs:
          type: integer
          minimum: 0
        steps:
          type: array
          maxItems: 128
          items:
            $ref: "#/components/schemas/OperationStep"
    ServiceRuntimeStatus:
      type: object
      additionalProperties: false
      required: [serviceId, serviceInstanceId, driver, mode, state, stateVersion, dependsOn, containers]
      properties:
        serviceId:
          $ref: "#/components/schemas/ManifestId"
        serviceInstanceId:
          type: string
          pattern: "^svi_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
        driver:
          type: string
          enum: [process, compose]
        mode:
          type: string
          enum: [daemon, oneshot]
        state:
          type: string
          enum: [stopped, waiting_dependency, starting, waiting_ready, ready, degraded, completed, stopping, failed, unknown]
        stateVersion:
          type: integer
          minimum: 1
        pid:
          type: integer
          minimum: 1
        processStartedAt:
          type: string
          format: date-time
        commandDigest:
          type: string
          pattern: "^[0-9a-f]{64}$"
        exitCode:
          type: integer
          minimum: 0
        dependsOn:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/ManifestId"
        containers:
          type: array
          maxItems: 64
          items:
            $ref: "#/components/schemas/ComposeContainerRuntimeStatus"
    ComposeContainerRuntimeStatus:
      type: object
      additionalProperties: false
      required: [service, state, health, exitCode]
      properties:
        service:
          type: string
          pattern: "^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$"
        state:
          type: string
          enum: [running, stopped, exited, restarting, paused, dead, created]
        health:
          type: string
          enum: [healthy, unhealthy, starting, ""]
        exitCode:
          type: integer
    PortRuntimeStatus:
      type: object
      additionalProperties: false
      required: [logicalName, port, source, replaced]
      properties:
        logicalName:
          $ref: "#/components/schemas/ManifestId"
        port:
          type: integer
          minimum: 1024
          maximum: 65535
        source:
          type: string
          enum: [request, workspace, sticky, preferred, fallback]
        replaced:
          type: boolean
        conflictPort:
          type: integer
          minimum: 1024
          maximum: 65535
    SystemRuntimeStatus:
      type: object
      additionalProperties: false
      required: [systemId, workspaceId, state, services, ports]
      properties:
        systemId:
          $ref: "#/components/schemas/ManifestId"
        workspaceId:
          $ref: "#/components/schemas/WorkspaceId"
        state:
          type: string
          enum: [stopping, failed, starting, degraded, running, stopped]
        instanceId:
          type: string
          pattern: "^si_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
        manifestDigest:
          type: string
          pattern: "^[0-9a-f]{64}$"
        resolvedSpecDigest:
          type: string
          pattern: "^[0-9a-f]{64}$"
        startedAt:
          type: string
          format: date-time
        stoppedAt:
          type: string
          format: date-time
        services:
          type: array
          items:
            $ref: "#/components/schemas/ServiceRuntimeStatus"
        ports:
          type: array
          items:
            $ref: "#/components/schemas/PortRuntimeStatus"
...
    ManifestId:
      type: string
      pattern: "^[a-z][a-z0-9-]{0,62}$"
...
    WorkspaceId:
      type: string
      pattern: "^ws_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
...
  responses:
...
    BadRequest:
      description: The request or manifest syntax is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: The request conflicts with current catalog state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    UnprocessableManifest:
      description: The manifest is structurally valid input but cannot be registered safely.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...